finish implementation of prop and method
